% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.r
\name{dif}
\alias{dif}
\title{dif: A custom R6 class}
\description{
dif: A custom R6 class

dif: A custom R6 class
}
\section{public methods}{

\describe{
  \item{\code{new()}}{Creates a new object of class `dif`.}
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dif-new}{\code{dif$new()}}
\item \href{#method-dif-c2a}{\code{dif$c2a()}}
\item \href{#method-dif-c_km}{\code{dif$c_km()}}
\item \href{#method-dif-s_m}{\code{dif$s_m()}}
\item \href{#method-dif-s_c}{\code{dif$s_c()}}
\item \href{#method-dif-s_c_inv}{\code{dif$s_c_inv()}}
\item \href{#method-dif-bt_o}{\code{dif$bt_o()}}
\item \href{#method-dif-tt_m}{\code{dif$tt_m()}}
\item \href{#method-dif-ec}{\code{dif$ec()}}
\item \href{#method-dif-dc}{\code{dif$dc()}}
\item \href{#method-dif-clone}{\code{dif$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dif-new"></a>}}
\if{latex}{\out{\hypertarget{method-dif-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dif$new(d)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dif-c2a"></a>}}
\if{latex}{\out{\hypertarget{method-dif-c2a}{}}}
\subsection{Method \code{c2a()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dif$c2a(c)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dif-c_km"></a>}}
\if{latex}{\out{\hypertarget{method-dif-c_km}{}}}
\subsection{Method \code{c_km()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dif$c_km(d)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dif-s_m"></a>}}
\if{latex}{\out{\hypertarget{method-dif-s_m}{}}}
\subsection{Method \code{s_m()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dif$s_m()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dif-s_c"></a>}}
\if{latex}{\out{\hypertarget{method-dif-s_c}{}}}
\subsection{Method \code{s_c()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dif$s_c(t, sm)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dif-s_c_inv"></a>}}
\if{latex}{\out{\hypertarget{method-dif-s_c_inv}{}}}
\subsection{Method \code{s_c_inv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dif$s_c_inv(t, sm)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dif-bt_o"></a>}}
\if{latex}{\out{\hypertarget{method-dif-bt_o}{}}}
\subsection{Method \code{bt_o()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dif$bt_o(t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dif-tt_m"></a>}}
\if{latex}{\out{\hypertarget{method-dif-tt_m}{}}}
\subsection{Method \code{tt_m()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dif$tt_m(t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dif-ec"></a>}}
\if{latex}{\out{\hypertarget{method-dif-ec}{}}}
\subsection{Method \code{ec()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dif$ec(t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dif-dc"></a>}}
\if{latex}{\out{\hypertarget{method-dif-dc}{}}}
\subsection{Method \code{dc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dif$dc(em)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dif-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dif-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dif$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
